Changed trigger polarity to active-low
Changing the trigger to active-low allows the effective usage of
the FPGA internal Pull-Ups in systems where some trigger-slaves
might come up when there is no master yet driving the bus to a
defined state. If no external pull-resistors are in place, the
bus state would essentially be random. By changing the bus to
active-low and eventually enabling the weak pull-ups on mini_exp1,
the trigger line will always be in a defined state whether or not
a master or external circuitry is present.